% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{alias_names}
\alias{alias_names}
\title{Get all existing alias names from a posterior database}
\usage{
alias_names(type, pdb = pdb_default())
}
\arguments{
\item{type}{type of alias names. Cureently only posteriors.}

\item{pdb}{a \code{pdb} object.}
}
\description{
Get all existing alias names from a posterior database
}
