% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{as.data.frame.pdb_posterior}
\alias{as.data.frame.pdb_posterior}
\alias{as.data.frame.pdb_posteriors}
\title{Coerce a [pdb_posterior] to a Data Frame}
\usage{
\method{as.data.frame}{pdb_posterior}(x, row.names = NULL,
  optional = FALSE, ...)

\method{as.data.frame}{pdb_posteriors}(x, row.names = NULL,
  optional = FALSE, ...)
}
\arguments{
\item{x}{a [pdb_posterior] object.}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{Not used.}

\item{...}{further arguments to \code{as.data.frame} for a list.}
}
\description{
Coerce a [pdb_posterior] to a Data Frame
}
\details{
The dataframe will consist of one row per keyword.
}
