% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.pdb_posteriors_info}
\alias{as.data.frame.pdb_posteriors_info}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{pdb_posteriors_info}(x, row.names = NULL,
  optional = FALSE, ...)
}
\arguments{
\item{x}{any R object.}

\item{optional}{Not used.}

\item{NULL}{or a character vector giving the row names for the data frame. Missing values are not allowed.}
}
\description{
Coerce to a Data Frame
}
