% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_checked_draws.R
\name{assert_checked_reference_posterior_draws}
\alias{assert_checked_reference_posterior_draws}
\alias{assert_checked_reference_posterior_draws.pdb_reference_posterior_draws}
\alias{assert_checked_reference_posterior_draws.pdb_reference_posterior_info}
\alias{assert_checked_summary_statistics_draws}
\alias{assert_checked_summary_statistics_draws.pdb_reference_posterior_draws}
\alias{assert_checked_summary_statistics_draws.pdb_reference_posterior_summary_statistic}
\alias{assert_checked_summary_statistics_draws.pdb_reference_posterior_info}
\title{Assert that all checks of the reference posterior draws
are true}
\usage{
assert_checked_reference_posterior_draws(x)

\method{assert_checked_reference_posterior_draws}{pdb_reference_posterior_draws}(x)

\method{assert_checked_reference_posterior_draws}{pdb_reference_posterior_info}(x)

assert_checked_summary_statistics_draws(x)

\method{assert_checked_summary_statistics_draws}{pdb_reference_posterior_draws}(x)

\method{assert_checked_summary_statistics_draws}{pdb_reference_posterior_summary_statistic}(x)

\method{assert_checked_summary_statistics_draws}{pdb_reference_posterior_info}(x)
}
\arguments{
\item{x}{a [pdb_reference_posterior_draws] object}
}
\description{
Assert that all checks of the reference posterior draws
are true
}
\details{
This functionasserts that the reference posterior draws comply
with the criterias of the reference posterior draws.

See \url{https://github.com/stan-dev/posteriordb/blob/master/doc/REFERENCE_POSTERIOR_DEFINITION.md} for details.
}
