% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_standard.R
\name{assert_gold_standard_info}
\alias{assert_gold_standard_info}
\title{Assert a gold_standard info}
\usage{
assert_gold_standard_info(x)
}
\arguments{
\item{x}{a an object to assert is a gold_standard_info object.}
}
\description{
Assert a gold_standard info
}
\keyword{internal}
