% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_standard.R
\name{assert_reference_posterior_info}
\alias{assert_reference_posterior_info}
\title{Assert a [pdb_reference_posterior_info] object}
\usage{
assert_reference_posterior_info(x)
}
\arguments{
\item{x}{a an object to assert is a [pdb_reference_posterior_info] object.}
}
\description{
Assert a [pdb_reference_posterior_info] object
}
\keyword{internal}
