% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{bibliography}
\alias{bibliography}
\alias{pdb_bibliography}
\title{Get the bibliography of a posterior database}
\usage{
bibliography(pdb, ...)

pdb_bibliography(pdb, ...)
}
\arguments{
\item{pdb}{a posterior database connection}

\item{...}{further arguments passed to [bibtex::read.bib]}
}
\description{
Get the bibliography of a posterior database
}
