% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_posterior.R
\name{check_pdb_posterior}
\alias{check_pdb_posterior}
\title{Test a posterior}
\usage{
check_pdb_posterior(po, run_stan_code_checks = TRUE, verbose = TRUE)
}
\arguments{
\item{po}{a [pdb_posterior] to check.}

\item{run_stan_code_checks}{should checks using Rstan be run?}

\item{verbose}{should check results be printed?}
}
\description{
Test a posterior
}
\details{
The function checks a posterior is consistent.
}
