% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_posterior.R
\name{check_posterior}
\alias{check_posterior}
\title{Test a posterior}
\usage{
check_posterior(po, run_stan_code_checks = TRUE)
}
\arguments{
\item{po}{a [pdb_posterior] to check.}

\item{run_stan_code_checks}{should checks using Rstan be run?}
}
\description{
Test a posterior
}
\details{
The function checks a posterior is consistent.
}
