% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reference_posterior.R
\name{check_reference_posterior_draws}
\alias{check_reference_posterior_draws}
\alias{check_reference_posterior_draws.character}
\alias{check_reference_posterior_draws.pdb_posterior}
\alias{check_reference_posterior_draws.pdb_reference_posterior_draws}
\title{Check that reference posterior draws follows the
reference posterior draws definition.}
\usage{
check_reference_posterior_draws(x, ...)

\method{check_reference_posterior_draws}{character}(x, ...)

\method{check_reference_posterior_draws}{pdb_posterior}(x, ...)

\method{check_reference_posterior_draws}{pdb_reference_posterior_draws}(x, ...)
}
\arguments{
\item{x}{a posterior name, posterior object or reference_posterior_draws object}

\item{...}{currently not used.}
}
\description{
Check that reference posterior draws follows the
reference posterior draws definition.
}
