% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/compute_reference_posterior_draws_stan_sampling.R
\name{compute_reference_posterior_draws}
\alias{compute_reference_posterior_draws}
\title{Compute a Reference Posteriors for a reference posterior info object}
\usage{
compute_reference_posterior_draws(rpi, pdb = pdb_default())
}
\arguments{
\item{rpi}{a [reference_posterior_info] object.}

\item{pdb}{a [pdb] object.}
}
\description{
Compute a Reference Posteriors for a reference posterior info object
}
