% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/compute_reference_posterior_draws_stan_sampling.R
\name{compute_reference_posterior_draws_stan_sampling}
\alias{compute_reference_posterior_draws_stan_sampling}
\title{Compute a Reference Posteriors using Rstan}
\usage{
compute_reference_posterior_draws_stan_sampling(rpi, pdb)
}
\arguments{
\item{rpi}{a [reference_posterior_info] object.}

\item{pdb}{a [pdb] object.}
}
\description{
Compute a Reference Posteriors using Rstan
}
