% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_summary_statistics.R
\name{compute_reference_posterior_summary_statistic}
\alias{compute_reference_posterior_summary_statistic}
\alias{compute_summary_statistic}
\title{Compute Summary statistics from a reference posterior}
\usage{
compute_reference_posterior_summary_statistic(rpd, summary_statistic = "mean")

compute_summary_statistic(rpd, summary_statistic = "mean")
}
\arguments{
\item{rpd}{a [reference_posterior_summary_statistic] object}

\item{summary_statistic}{summary statistic to compute}
}
\description{
Compute Summary statistics from a reference posterior
}
