% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{data_names}
\alias{data_names}
\title{Get all existing data names from a posterior database}
\usage{
data_names(pdb)
}
\arguments{
\item{pdb}{a \code{pdb} object.}
}
\description{
Get all existing data names from a posterior database
}
