% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{Posterior Data Sets}
\usage{
dataset(x, ...)
}
\arguments{
\item{x}{a \code{posterior} object to access dataset from.}

\item{...}{Currently not used.}
}
\description{
Posterior Data Sets
}
