% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{essb}
\alias{essb}
\alias{esst}
\title{Simulated values for effective sample size (tail and bulk)}
\format{An object of class \code{numeric} of length 100000.}
\usage{
`data("essb")`

`data("esst")`
}
\description{
A simulated NULL distribution for ESS tail (esst) and ESS bulk (essb).

The variance in ESS can be quite large so to test that a gold standard
posterior contains roughly indpendent draws, we simulate perfectly
independent draws and compute the ESS distribution.

This is mainly used to test gold standard posteriors-

See `data-raw/ess.R` for simulation code.
}
\keyword{datasets}
