% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_code.R
\name{framework}
\alias{framework}
\alias{framework.pdb_model_code}
\alias{framework<-}
\alias{framework<-.character}
\alias{framework<-.pdb_model_code}
\title{Identify the framework for a given [model_code]}
\usage{
framework(x)

\method{framework}{pdb_model_code}(x)

framework(x) <- value

\method{framework}{character}(x) <- value

\method{framework}{pdb_model_code}(x) <- value
}
\arguments{
\item{x}{a [pdb_model_code] object.}

\item{value}{a supported framework.}
}
\description{
Identify the framework for a given [model_code]
}
