% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data}
\alias{get_data}
\alias{get_data.pdb_posterior}
\alias{get_data.character}
\alias{get_data.pdb_data_info}
\alias{get_data.list}
\alias{pdb_data}
\title{Posterior Data Sets}
\usage{
get_data(x, ...)

\method{get_data}{pdb_posterior}(x, ...)

\method{get_data}{character}(x, pdb = pdb_default(), ...)

\method{get_data}{pdb_data_info}(x, pdb = pdb_default(), ...)

\method{get_data}{list}(x, info, ...)

pdb_data(x, ...)
}
\arguments{
\item{x}{an object to access data from.}

\item{...}{Currently not used.}

\item{pdb}{a \code{pdb} object.}

\item{info}{a \code{pdb_data_info} object.}
}
\description{
Posterior Data Sets
}
