% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_github.R
\name{github_pat}
\alias{github_pat}
\title{Retrieve Github personal access token.}
\usage{
github_pat(pdb = NULL)
}
\arguments{
\item{pdb}{A posterior datasbase object to extract pat from.}
}
\description{
A github personal access token
Looks in env var `GITHUB_PAT`
}
