% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_standard.R
\name{gold_standard_draws}
\alias{gold_standard_draws}
\alias{gold_standard_draws.character}
\alias{gold_standard_draws.pdb_posterior}
\alias{gold_standard_draws.pdb_gold_standard_info}
\alias{gold_standard_draws.stanfit}
\alias{assert_gold_standard_draws}
\title{Gold standard posterior draws}
\usage{
gold_standard_draws(x, ...)

\method{gold_standard_draws}{character}(x, pdb = pdb_default(), ...)

\method{gold_standard_draws}{pdb_posterior}(x, ...)

\method{gold_standard_draws}{pdb_gold_standard_info}(x, pdb = pdb_default(), ...)

\method{gold_standard_draws}{stanfit}(x, pdb = pdb_default(), ...)

assert_gold_standard_draws(x)
}
\arguments{
\item{x}{a [posterior] object or a posterior name.}

\item{...}{further arguments supplied to specific methods.}

\item{pdb}{a [pdb] object (if [x] is a posterior name)}
}
\value{
a gold_standard, draws_list object.
}
\description{
Gold standard posterior draws
}
