% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_standard.R
\name{gold_standard_file_path}
\alias{gold_standard_file_path}
\title{Extract data for posterior}
\usage{
gold_standard_file_path(x)
}
\arguments{
\item{x}{a \code{posterior} object.}
}
\description{
Extract data for posterior
}
