% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_standard.R
\name{gold_standard_info}
\alias{gold_standard_info}
\alias{gold_standard_info.pdb_posterior}
\alias{gold_standard_info.character}
\title{Gold standard information}
\usage{
gold_standard_info(x, pdb = pdb_default(), ...)

\method{gold_standard_info}{pdb_posterior}(x, ...)

\method{gold_standard_info}{character}(x, pdb = pdb_default(), ...)
}
\arguments{
\item{x}{a \code{posterior} object to access the gold standard for.}

\item{pdb}{a \code{pdb} posterior database connection.}

\item{...}{Currently not used.}
}
\description{
Gold standard information
}
\details{
Gold standard information contain information on how, when and by whom
the gold standard was computed.
}
\seealso{
gold_standard_draws
}
