% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{handle_aliases}
\alias{handle_aliases}
\title{Return the name that the alias points to}
\usage{
handle_aliases(x, type, pdb)
}
\arguments{
\item{x}{a name to lookup in alias list}

\item{type}{type of alias names. Cureently only posteriors.}

\item{pdb}{a \code{pdb} object.}
}
\description{
Return the name that the alias points to
}
