% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{info<-}
\alias{info<-}
\title{Set information to a pdb object}
\usage{
info(x) <- value
}
\arguments{
\item{x}{a [pdb_data], [pdb_model_code] and [pdb_reference_posterior]}

\item{value}{an info object}
}
\description{
Set information to a pdb object
}
