% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{info}
\alias{info}
\title{Extract information from a pdb object}
\usage{
info(x)
}
\arguments{
\item{x}{a [pdb_data], [pdb_model_code] and [pdb_reference_posterior]}

\item{info}{a [pdb_]}
}
\description{
Extract information from a pdb object
}
