% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_code.R
\name{model_code_file_paths}
\alias{model_code_file_paths}
\title{Extract file paths to model codes from posterior object}
\usage{
model_code_file_paths(x)
}
\arguments{
\item{x}{a \code{posterior} object.}
}
\value{
a list with paths to different model codes
}
\description{
Extract file paths to model codes from posterior object
}
