% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{pdb<-}
\alias{pdb<-}
\title{Set pdb slot}
\usage{
pdb(x) <- value
}
\arguments{
\item{x}{a [pdb_data], [pdb_model_code] and [pdb_reference_posterior]}

\item{value}{a pdb object}
}
\description{
Set pdb slot
}
