% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{pdb}
\alias{pdb}
\title{Posterior database (pdb) constructor}
\usage{
pdb(x = getwd())
}
\arguments{
\item{x}{a path to the posterior database
(i.e. where folders content and posteriors exists)}
}
\value{
a \code{pdb} object
}
\description{
Posterior database (pdb) constructor
}
