% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{pdb_aliases}
\alias{pdb_aliases}
\title{Extract aliases list from posterior database}
\usage{
pdb_aliases(type, pdb)
}
\arguments{
\item{type}{type of alias names. Cureently only posteriors.}

\item{pdb}{a \code{pdb} object.}
}
\description{
Extract aliases list from posterior database
}
