% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{pdb_cache_path}
\alias{pdb_cache_path}
\title{Returns a writable cache path for a pdb and a path
It will create the directory if it does not exist.}
\usage{
pdb_cache_path(pdb, path)
}
\arguments{
\item{pdb}{a \code{pdb} object.}

\item{path}{a \code{pdb} path.}
}
\description{
Returns a writable cache path for a pdb and a path
It will create the directory if it does not exist.
}
