% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{pdb_cache_rm}
\alias{pdb_cache_rm}
\title{Remove object from cache}
\usage{
pdb_cache_rm(x, pdb = pdb_default())
}
\arguments{
\item{x}{an object to remove from cache}
}
\description{
Remove object from cache
}
