% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{pdb_cached_local_file_path}
\alias{pdb_cached_local_file_path}
\title{Read json file from \code{path}}
\usage{
pdb_cached_local_file_path(pdb, path, unzip = FALSE, ...)
}
\arguments{
\item{pdb}{a \code{pdb} to read from.}

\item{path}{a \code{pdb} to read from.}

\item{unzip}{if true, path is zipped and should be unzipped to cache.}
}
\description{
Read json file from \code{path}
}
\details{
Copies the file to the cache and return path
}
