% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{pdb_clear_cache}
\alias{pdb_clear_cache}
\title{Clear posterior database cache}
\usage{
pdb_clear_cache(pdb)
}
\arguments{
\item{pdb}{a \code{pdb} to clear cache for}
}
\description{
Clear posterior database cache
}
\keyword{internal}
