% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{pdb_file_copy}
\alias{pdb_file_copy}
\title{Copy a file from a pdb to a local path}
\usage{
pdb_file_copy(pdb, from, to, overwrite = FALSE, ...)
}
\arguments{
\item{from}{a path in the pdb}

\item{to}{a local file path}
}
\value{
a boolean indicator as file.copy indicating success.
}
\description{
Copy a file from a pdb to a local path
}
