% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{pdb_local}
\alias{pdb_local}
\alias{pdb}
\title{Create a local Posterior Database (pdb) connections}
\usage{
pdb_local(path = getOption("pdb_path", file.path(getwd(),
  "posterior_database")), cache_path = tempdir())

pdb(pdb_id, pdb_type = "local", cache_path = tempdir(), ...)
}
\arguments{
\item{path}{a local path to the posterior db}

\item{cache_path}{The path to the pdb cache. Default is R temporary directory.
This is used to store files locally and without affecting the database.}

\item{pdb_id}{how to identify the pdb (path for local pdb, repo for github pdb)}

\item{repo}{a github repository with a posterior db}
}
\value{
a \code{pdb} object
}
\description{
Create a local Posterior Database (pdb) connections
}
