% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R, R/pdb_github.R
\name{pdb_version}
\alias{pdb_version}
\alias{pdb_version.pdb_local}
\alias{pdb_version.pdb_github}
\title{Get version of the \code{pdb}}
\usage{
pdb_version(pdb, ...)

\method{pdb_version}{pdb_local}(pdb, ...)

\method{pdb_version}{pdb_github}(pdb, ...)
}
\arguments{
\item{pdb}{a \code{pdb} object to return version for.}
}
\value{
the git sha for the posterior database.
}
\description{
Get version of the \code{pdb}
}
