% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\title{Access a posterior in the posterior database}
\usage{
posterior(name, pdb = pdb())
}
\arguments{
\item{name}{a posterior name that exist in the posterior database}

\item{pdb}{a \code{pdb} posterior database object.}
}
\description{
Access a posterior in the posterior database
}
