% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_data.R
\name{posterior_data_file_path}
\alias{posterior_data_file_path}
\alias{posterior_data}
\title{Extract data for posterior}
\usage{
posterior_data_file_path(x)

posterior_data(x)
}
\arguments{
\item{x}{a \code{posterior} object.}
}
\description{
Extract data for posterior
}
