% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_fit.R
\name{posterior_draws}
\alias{posterior_draws}
\alias{posterior_draws.pdb_posterior_fit}
\title{Extract posterior draws}
\usage{
posterior_draws(x, ...)

\method{posterior_draws}{pdb_posterior_fit}(x, ...)
}
\arguments{
\item{x}{An object to extract draws from.}

\item{...}{Other arguments passed to or from other methods.}
}
\description{
Extract posterior draws
}
