% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_fit.R
\name{posterior_fit}
\alias{posterior_fit}
\title{Constructor for posterior_fit objects}
\usage{
posterior_fit(x)
}
\arguments{
\item{x}{an object to convert to a \code{posterior_fit}.}
}
\description{
Constructor for posterior_fit objects
}
