% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{posteriors_tbl_df}
\alias{posteriors_tbl_df}
\title{Create Posterior Database tibbles}
\usage{
posteriors_tbl_df(pdb, ...)
}
\arguments{
\item{pdb}{a \code{pdb} object.}

\item{...}{further arguments to supply to \code{dplyr::filter()}}
}
\description{
Create Posterior Database tibbles
}
\details{
Generates a tibble data frame of all posteriors
in the database and their side information
}
