% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{read_json_from_pdb}
\alias{read_json_from_pdb}
\title{Read JSON objects from the posterior database}
\usage{
read_json_from_pdb(fn, path, pdb, ...)
}
\arguments{
\item{fn}{file name}

\item{path}{path to file name in pdb}

\item{pdb}{a [pdb] object}

\item{...}{further arguments supplied to [jsonlite::read_json()]}
}
\description{
Read JSON objects from the posterior database
}
