% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{read_posterior}
\alias{read_posterior}
\alias{read_pdb_posterior}
\title{read a posterior object from the data base}
\usage{
read_posterior(name, pdb)

read_pdb_posterior(name, pdb)
}
\description{
read a posterior object from the data base
}
