% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_posterior.R, R/summary_statistic.R
\name{reference_posterior_draws_file_path}
\alias{reference_posterior_draws_file_path}
\alias{reference_posterior_draws_file_path.pdb_posterior}
\alias{reference_posterior_draws_file_path.character}
\alias{reference_posterior_draws_file_path.pdb_reference_posterior_info}
\title{Extract data for posterior}
\usage{
reference_posterior_draws_file_path(x, ...)

\method{reference_posterior_draws_file_path}{pdb_posterior}(x, ...)

\method{reference_posterior_draws_file_path}{character}(x, pdb = pdb_default(), ...)

\method{reference_posterior_draws_file_path}{pdb_reference_posterior_info}(x, pdb = pdb_default(), ...)

reference_posterior_draws_file_path(x, ...)

\method{reference_posterior_draws_file_path}{pdb_posterior}(x, ...)

\method{reference_posterior_draws_file_path}{character}(x, pdb = pdb_default(), ...)

\method{reference_posterior_draws_file_path}{pdb_reference_posterior_info}(x, pdb = pdb_default(), ...)
}
\arguments{
\item{x}{an object to access file path to.}

\item{...}{further arguments supplied to methods.}

\item{pdb}{a \code{pdb} object.}
}
\description{
Extract data for posterior

Extract data for posterior
}
