% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_standard.R
\name{reference_posterior_info}
\alias{reference_posterior_info}
\alias{reference_posterior_info.pdb_posterior}
\alias{reference_posterior_info.character}
\title{Reference posterior information}
\usage{
reference_posterior_info(x, pdb = pdb_default(), ...)

\method{reference_posterior_info}{pdb_posterior}(x, ...)

\method{reference_posterior_info}{character}(x, pdb = pdb_default(), type = ...)
}
\arguments{
\item{x}{a \code{posterior} object to access the reference posterior for.}

\item{pdb}{a \code{pdb} posterior database connection.}

\item{...}{Currently not used.}
}
\description{
Reference posterior information
}
\details{
Reference posterior information on how, when and by whom
the reference posterior was computed.
}
\seealso{
reference_posterior_draws
}
