% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{reference_posterior_names}
\alias{reference_posterior_names}
\alias{reference_posterior_names.pdb_local}
\title{Get all existing reference posterior names from a posterior database}
\usage{
reference_posterior_names(pdb = pdb_default(), type, ...)

\method{reference_posterior_names}{pdb_local}(pdb = pdb_default(), type, ...)
}
\arguments{
\item{pdb}{a \code{pdb} object.}

\item{...}{Further argument to methods.}
}
\description{
Get all existing reference posterior names from a posterior database
}
