% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_pdb.R
\name{remove_from_pdb}
\alias{remove_from_pdb}
\alias{remove_posteriors_from_pdb}
\alias{remove_data_from_pdb}
\alias{remove_models_from_pdb}
\alias{remove_reference_posterior_from_pdb}
\title{Remove objects from local pdb}
\usage{
remove_from_pdb(x, type, pdb)

remove_posteriors_from_pdb(x, pdb)

remove_data_from_pdb(x, pdb)

remove_models_from_pdb(x, pdb)

remove_reference_posterior_from_pdb(x, pdb)
}
\arguments{
\item{x}{an object name to remove to the pdb.}

\item{type}{Type of object to remove.}

\item{pdb}{the pdb to remove from. Only a local pdb.}
}
\description{
a function to simplify writing to a local pdb.
}
