% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pdb.R
\name{remove_pdb}
\alias{remove_pdb}
\alias{remove_pdb.pdb_data}
\alias{remove_pdb.pdb_data_info}
\alias{remove_pdb.character}
\title{Remove objects from local pdb}
\usage{
remove_pdb(x, pdb, ...)

\method{remove_pdb}{pdb_data}(x, pdb, ...)

\method{remove_pdb}{pdb_data_info}(x, pdb, ...)

\method{remove_pdb}{character}(x, pdb, ...)
}
\arguments{
\item{x}{an object to remove to the pdb.}

\item{pdb}{the pdb to remove from. Currently only a local pdb.}

\item{...}{further arguments supplied to methods.}
}
\description{
a function to simplify removing object from a local pdb.
}
