% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jags.R
\name{run_jags}
\alias{run_jags}
\alias{run_jags.pdb_posterior}
\title{Run JAGS on a posterior}
\usage{
run_jags(x, args.jags.model, args.coda.samples, ...)

\method{run_jags}{pdb_posterior}(x, args.jags.model, args.coda.samples, ...)
}
\arguments{
\item{x}{a [pdb_posterior] object.}

\item{args.jags.model}{Arguments as a list supplied to [rjags::jags.model] (excluding file and data)}

\item{args.coda.samples}{Arguments as a list supplied to [rjags::coda.samples] (excluding model)}

\item{...}{currently not in use.}
}
\description{
Run JAGS on a posterior
}
\references{
See \url{http://mcmc-jags.sourceforge.net/}.
}
