% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stan.R
\name{run_stan}
\alias{run_stan}
\alias{run_stan.pdb_posterior}
\title{Run Stan on a posterior}
\usage{
run_stan(x, stan_args, ...)

\method{run_stan}{pdb_posterior}(x, stan_args, ...)
}
\arguments{
\item{x}{a [pdb_posterior] object.}

\item{stan_args}{Arguments supplied to [rstan::stan] to compute the posterior}
}
\description{
Run Stan on a posterior
}
