% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{setup_pdb}
\alias{setup_pdb}
\alias{setup_pdb.pdb_local}
\title{Setup object specific part of pdb object}
\usage{
setup_pdb(pdb, ...)

\method{setup_pdb}{pdb_local}(pdb, ...)
}
\arguments{
\item{pdb}{a \code{pdb} object.}

\item{...}{further arguments supplied to specific methods (not in use)}
}
\description{
Setup object specific part of pdb object
}
\keyword{internal}
