% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_code.R
\name{stan_code_file_path}
\alias{stan_code_file_path}
\alias{stan_code}
\title{Extract stan code file path or stan code from posterior object}
\usage{
stan_code_file_path(x)

stan_code(x)
}
\arguments{
\item{x}{a \code{posterior} object.}
}
\description{
Extract stan code file path or stan code from posterior object
}
