% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_data.R
\name{stan_data_file_path}
\alias{stan_data_file_path}
\alias{stan_data}
\title{Extract data for posterior}
\usage{
stan_data_file_path(x)

stan_data(x)
}
\arguments{
\item{x}{a \code{posterior} object.}
}
\description{
Extract data for posterior
}
