% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stancode.R
\name{stancode_file_path}
\alias{stancode_file_path}
\alias{stancode}
\title{Extract stan code file path from posterior}
\usage{
stancode_file_path(x)

stancode(x)
}
\arguments{
\item{x}{a \code{posterior} object.}
}
\description{
Extract stan code file path from posterior
}
