% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_standard.R
\name{subset.pdb_reference_posterior_draws}
\alias{subset.pdb_reference_posterior_draws}
\title{Subset a [pdb_reference_posterior_draws] object}
\usage{
\method{subset}{pdb_reference_posterior_draws}(x, variable, ...)
}
\arguments{
\item{x}{a [pdb_reference_posterior_draws] to subest}

\item{variable}{parameter names to subset.}

\item{...}{Further arguments (not used).}
}
\description{
Subset a [pdb_reference_posterior_draws] object
}
