% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistic.R
\name{summary_statistic_type}
\alias{summary_statistic_type}
\title{Extract the type of summary statistic}
\usage{
summary_statistic_type(x)
}
\arguments{
\item{x}{a [pdb_reference_posterior_summary_statistic]}
}
\description{
Extract the type of summary statistic
}
