% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gold_standard.R
\name{test_gold_standard_draws}
\alias{test_gold_standard_draws}
\alias{test_gold_standard_draws.character}
\alias{test_gold_standard_draws.pdb_posterior}
\alias{test_gold_standard_draws.pdb_gold_standard_draws}
\title{Test that gold standard draws follows the
gold standard definition.}
\usage{
test_gold_standard_draws(x, ...)

\method{test_gold_standard_draws}{character}(x, ...)

\method{test_gold_standard_draws}{pdb_posterior}(x, ...)

\method{test_gold_standard_draws}{pdb_gold_standard_draws}(x, ...)
}
\arguments{
\item{x}{a posterior name, posterior object or gold_standard_draws object}

\item{...}{currently not used.}
}
\description{
Test that gold standard draws follows the
gold standard definition.
}
