% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_posterior.R
\name{test_posterior}
\alias{test_posterior}
\title{Test a posterior}
\usage{
test_posterior(po)
}
\arguments{
\item{po}{a [pdb_posterior] to check.}
}
\description{
Test a posterior
}
\details{
The function checks a posterior is consistent.
}
