% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reference_posterior.R
\name{thin_draws.pdb_reference_posterior_draws}
\alias{thin_draws.pdb_reference_posterior_draws}
\title{Thin draws to reduce their size and autocorrelation of the chains.}
\usage{
thin_draws.pdb_reference_posterior_draws(x, thin, ...)
}
\arguments{
\item{x}{An R object for which the methods are defined.}

\item{thin}{A positive integer specifying the period for selecting draws.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\value{
A thinned [pdb_reference_posterior_draws] object.
}
\description{
Thin [pdb_reference_posterior_draws] objects to reduce their size and autocorrelation of the chains.
}
