% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pdb.R
\name{check_pdb}
\alias{check_pdb}
\alias{check_pdb_run_stan}
\alias{check_pdb_stan_syntax}
\alias{check_pdb_references}
\title{Check the content a posterior database}
\usage{
check_pdb(pdb, posterior_idx = NULL)

check_pdb_run_stan(pdb, posterior_idx = NULL)

check_pdb_stan_syntax(pdb, posterior_idx = NULL)

check_pdb_references(pdb, posterior_idx = NULL)
}
\arguments{
\item{pdb}{a \code{pdb} object}

\item{posterior_idx}{an index vector indicating what posteriors to check.}
}
\value{
a boolean indicating if the pdb works as it should.
}
\description{
Check the content a posterior database
}
\details{
[check_pdb()] checks that the content exists as specified
[check_pdb_run_stan()] test to run all posteriors with stan models.
[check_pdb_stan_syntax()] check that all stan model code files can be parsed.
}
