% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R, R/pdb_github.R
\name{data_names}
\alias{data_names}
\alias{data_names.pdb_local}
\alias{data_names.pdb_github}
\alias{model_names.pdb_github}
\title{Get all existing data names from a posterior database}
\usage{
data_names(pdb, ...)

\method{data_names}{pdb_local}(pdb, ...)

\method{data_names}{pdb_github}(pdb, ...)

\method{model_names}{pdb_github}(pdb, ...)
}
\arguments{
\item{pdb}{a \code{pdb} object.}

\item{...}{Further argument to methods.}
}
\description{
Get all existing data names from a posterior database
}
