% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_names.R
\name{dataset_names}
\alias{dataset_names}
\title{Get all existing dataset names from a posterior database}
\usage{
dataset_names(pdbo)
}
\arguments{
\item{pdbo}{a \code{pdb} object.}
}
\description{
Get all existing dataset names from a posterior database
}
