% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gold_standard.R
\name{ess_bounds}
\alias{ess_bounds}
\title{Compute ESS tail and bulk bounds}
\usage{
ess_bounds(x, p = 0.95)
}
\arguments{
\item{x}{a [pdb_gold_standard_draws] object.}

\item{p}{the joint probability for independent ESS estimators.}
}
\description{
Compute ESS tail and bulk bounds
}
\details{
Compute the bounds for ESS tail and bulk based on 100 000
simulated ESS computations.
}
\keyword{internal}
