% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_github.R
\name{github_ref}
\alias{github_ref}
\title{Retrieve Github reference.}
\usage{
github_ref(pdb = NULL)
}
\arguments{
\item{pdb}{A posterior datasbase object to extract pat from.}
}
\description{
A github reference to use
Looks in env var `GITHUB_REF`
}
