% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_standard.R
\name{gold_standard}
\alias{gold_standard}
\title{Gold standard posterior draws}
\usage{
gold_standard(x, ...)
}
\arguments{
\item{x}{a \code{posterior} object to access the gold standard for.}

\item{...}{Currently not used.}
}
\description{
Gold standard posterior draws
}
