% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_standard.R
\name{gold_standard_draws_file_path}
\alias{gold_standard_draws_file_path}
\alias{gold_standard_draws_file_path.pdb_posterior}
\alias{gold_standard_draws_file_path.character}
\alias{gold_standard_draws_file_path.pdb_gold_standard_info}
\title{Extract data for posterior}
\usage{
gold_standard_draws_file_path(x, ...)

\method{gold_standard_draws_file_path}{pdb_posterior}(x, ...)

\method{gold_standard_draws_file_path}{character}(x, pdb = pdb_default(), ...)

\method{gold_standard_draws_file_path}{pdb_gold_standard_info}(x, pdb = pdb_default(), ...)
}
\arguments{
\item{x}{an object to access file path to.}

\item{...}{further arguments supplied to methods.}

\item{pdb}{a \code{pdb} object.}
}
\description{
Extract data for posterior
}
