% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_standard.R
\name{gold_standard_file_path}
\alias{gold_standard_file_path}
\title{Extract data for posterior}
\usage{
gold_standard_file_path(x, tempdir = TRUE)
}
\arguments{
\item{x}{a \code{posterior} object.}

\item{tempdir}{Should the file be copied to R temporary directory? Default is \code{TRUE}.}
}
\description{
Extract data for posterior
}
