% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{info}
\alias{info}
\alias{info<-}
\title{Extract information}
\usage{
info(x)

info(x) <- value
}
\arguments{
\item{x}{a [pdb_data], [pdb_model_code] and [pdb_reference_posterior]}
}
\description{
Extract information
}
