% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_code.R
\name{model_code_file_path}
\alias{model_code_file_path}
\alias{model_code}
\alias{stan_code_file_path}
\alias{stan_code}
\title{Extract file paths to model code from posterior object}
\usage{
model_code_file_path(x, framework, tempdir = TRUE)

model_code(x, framework)

stan_code_file_path(x, tempdir = TRUE)

stan_code(x)
}
\arguments{
\item{x}{a \code{posterior} object.}

\item{framework}{model code framework (e.g. \code{stan}).}

\item{tempdir}{Should the file be copied to R temporary directory? Default is \code{TRUE}.}
}
\description{
Extract file paths to model code from posterior object
}
