% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_code.R
\name{model_code_file_path}
\alias{model_code_file_path}
\alias{model_code}
\alias{pdb_model_code}
\alias{stan_code_file_path}
\alias{stan_code}
\alias{pdb_stan_code}
\title{Extract file paths to model code from posterior object}
\usage{
model_code_file_path(x, framework)

model_code(x, framework)

pdb_model_code(x, framework)

stan_code_file_path(x)

stan_code(x)

pdb_stan_code(x)
}
\arguments{
\item{x}{a \code{posterior} object.}

\item{framework}{model code framework (e.g. \code{stan}).}
}
\description{
Extract file paths to model code from posterior object
}
