% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R, R/model_info.R
\name{data_info}
\alias{data_info}
\alias{model_info}
\title{Access data and model information}
\usage{
data_info(po)

model_info(po)
}
\arguments{
\item{po}{a \code{posterior} object.}
}
\description{
Access data and model information
}
