% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{model_names}
\alias{model_names}
\title{Get all existing model names from a posterior database}
\usage{
model_names(pdb)
}
\arguments{
\item{pdb}{a \code{pdb} object.}
}
\description{
Get all existing model names from a posterior database
}
