% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{pdb_assert_file_exist}
\alias{pdb_assert_file_exist}
\alias{pdb_assert_file_exist.pdb_local}
\title{Assert that a file exists}
\usage{
pdb_assert_file_exist(pdb, path, ...)

\method{pdb_assert_file_exist}{pdb_local}(pdb, path, ...)
}
\arguments{
\item{pdb}{a \code{pdb} object.}

\item{path}{a \code{pdb} path.}

\item{...}{further arguments supplied to methods.}
}
\description{
Assert that a file exists
}
