% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_filter.R
\name{pdb_filter}
\alias{pdb_filter}
\title{Internal filter function}
\usage{
pdb_filter(pdb, path, ...)
}
\arguments{
\item{pdb}{a pdb connection}
}
\description{
Works for filtering models, data and posteriors.
}
\keyword{internal}
