% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\alias{posterior.character}
\alias{posterior.list}
\alias{pdb_posterior}
\title{Access a posterior in the posterior database}
\usage{
posterior(x, pdb = pdb_default(), ...)

\method{posterior}{character}(x, pdb = pdb_default(), ...)

\method{posterior}{list}(x, pdb = pdb_default(), ...)

pdb_posterior(x, pdb = pdb_default(), ...)
}
\arguments{
\item{x}{a posterior name that exist in the posterior database}

\item{pdb}{a \code{pdb} posterior database object.}

\item{...}{currently not in use.}
}
\description{
Access a posterior in the posterior database
}
