% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\title{Access a posterior in the posterior database}
\usage{
posterior(x)
}
\arguments{
\item{x}{a file path to a posterior json, or a posterior_name (require i.e. internet access)}
}
\description{
Access a posterior in the posterior database
}
