% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{posterior_dimension_names}
\alias{posterior_dimension_names}
\title{Construct dimension names from a posterior dimension list}
\usage{
posterior_dimension_names(x)
}
\arguments{
\item{x}{a dimensions slot from a [pdb_posterior]}
}
\description{
Construct dimension names from a posterior dimension list
}
