% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R, R/posterior_draws.R
\name{posterior_draws}
\alias{posterior_draws}
\alias{posterior_draws.posterior_fit}
\title{Extract posterior draws}
\usage{
posterior_draws(x)

\method{posterior_draws}{posterior_fit}(x)

posterior_draws(x)
}
\arguments{
\item{x}{an object to extract draws from.}

\item{x}{an object to extract draws from.}
}
\description{
Extract posterior draws

Extract posterior draws
}
