% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_fit.R
\name{posterior_fit}
\alias{posterior_fit}
\alias{pdb_posterior_fit}
\title{Constructor for \code{pdb_posterior_fit} objects}
\usage{
posterior_fit(path)

pdb_posterior_fit(path)
}
\arguments{
\item{path}{A file path to a JSON file containing a posterior fit object}
}
\description{
Constructor for \code{pdb_posterior_fit} objects
}
