% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R, R/pdb_github.R
\name{posterior_names}
\alias{posterior_names}
\alias{posterior_names.pdb_github}
\title{Get all existing posterior names from a posterior database.}
\usage{
posterior_names(pdb = pdb_default(), ...)

\method{posterior_names}{pdb_github}(pdb, ...)
}
\arguments{
\item{pdb}{a \code{pdb} object.}

\item{...}{further arguments supplied to specific methods (not in use)}
}
\description{
Get all existing posterior names from a posterior database.
}
