% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_names.R
\name{posterior_names}
\alias{posterior_names}
\title{Get all existing posterior names from a posterior database}
\usage{
posterior_names(pdbo)
}
\arguments{
\item{pdbo}{a \code{pdb} object.}
}
\description{
Get all existing posterior names from a posterior database
}
