% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_standard.R
\name{reference_posterior_info}
\alias{reference_posterior_info}
\alias{pdb_reference_posterior_info}
\alias{reference_posterior_draws_info}
\alias{pdb_reference_posterior_draws_info}
\alias{reference_posterior_expectations_info}
\alias{reference_posterior_info.pdb_posterior}
\alias{reference_posterior_info.character}
\alias{reference_posterior_info.list}
\title{Reference posterior information}
\usage{
reference_posterior_info(x, type, pdb = pdb_default(), ...)

pdb_reference_posterior_info(x, type, pdb = pdb_default(), ...)

reference_posterior_draws_info(x, pdb = pdb_default(), ...)

pdb_reference_posterior_draws_info(x, pdb = pdb_default(), ...)

reference_posterior_expectations_info(x, pdb = pdb_default(), ...)

\method{reference_posterior_info}{pdb_posterior}(x, type, ...)

\method{reference_posterior_info}{character}(x, type, pdb = pdb_default(), ...)

\method{reference_posterior_info}{list}(x, type = NULL, pdb = NULL, ...)
}
\arguments{
\item{x}{a \code{posterior} object to access the reference posterior for.}

\item{type}{Type of reference posterior [draws] or [expectations].}

\item{pdb}{a \code{pdb} posterior database connection.}

\item{...}{Currently not used.}
}
\description{
Reference posterior information
}
\details{
Reference posterior information on how, when and by whom
the reference posterior was computed.
}
\seealso{
reference_posterior_draws
}
