% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_standard.R
\name{subset.pdb_gold_standard_draws}
\alias{subset.pdb_gold_standard_draws}
\title{Subset a [pdb_gold_standard_draws] object}
\usage{
\method{subset}{pdb_gold_standard_draws}(x, variable, ...)
}
\arguments{
\item{x}{a [pdb_gold_standard_draws] to subest}

\item{variable}{Parameters to subset.}

\item{...}{Further arguments (not used).}
}
\description{
Subset a [pdb_gold_standard_draws] object
}
