% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reference_posterior.R
\name{test_reference_posterior_draws}
\alias{test_reference_posterior_draws}
\alias{test_reference_posterior_draws.character}
\alias{test_reference_posterior_draws.pdb_posterior}
\alias{test_reference_posterior_draws.pdb_reference_posterior_draws}
\title{Test that gold standard draws follows the
gold standard definition.}
\usage{
test_reference_posterior_draws(x, ...)

\method{test_reference_posterior_draws}{character}(x, ...)

\method{test_reference_posterior_draws}{pdb_posterior}(x, ...)

\method{test_reference_posterior_draws}{pdb_reference_posterior_draws}(x, ...)
}
\arguments{
\item{x}{a posterior name, posterior object or reference_posterior_draws object}

\item{...}{currently not used.}
}
\description{
Test that gold standard draws follows the
gold standard definition.
}
